Improving logging: adding proper logging for exceptions that can appear when the user uses incorrect type for deserialization. (#61)

+++ ktoml-core/src/commonTest/kotlin/decoder/ArrayDecoderTest.kt
@@ -9,10 +9,10 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 @Serializable
-data class SimpleArray(val a: List<Int>)
+data class SimpleArray(val a: List<Long>)
 
 @Serializable
-data class NestedArray(val a: List<List<Int>>)
+data class NestedArray(val a: List<List<Long>>)
 
 @Serializable
 data class ArrayInInlineTable(val table: InlineTable)
